Revert "AbstractCache"
This reverts commit a02842248e5dc123e3c3353375476f9b658d58a7.
