+++ src/ITensorNetworks.jl
@@ -28,10 +28,9 @@ include("edge_sequences.jl")
 include("formnetworks/abstractformnetwork.jl")
 include("formnetworks/bilinearformnetwork.jl")
 include("formnetworks/quadraticformnetwork.jl")
-include("caches/abstractbeliefpropagationcache.jl")
 include("caches/beliefpropagationcache.jl")
-#include("caches/boundarympscacheutils.jl")
-#include("caches/boundarympscache.jl")
+include("caches/boundarympscacheutils.jl")
+include("caches/boundarympscache.jl")
 include("contraction_tree_to_graph.jl")
 include("gauging.jl")
 include("utils.jl")
